// frontend/viewer/src/lib/dotnet-types/generated-types/FwLiteShared/Events/IFwEvent.ts
/* eslint-disable */
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import type {FwEventType} from './FwEventType';

export interface IFwEvent
{
	type: FwEventType;
	isGlobal: boolean;
}
/* eslint-enable */