// frontend/viewer/src/lib/dotnet-types/generated-types/FwLiteShared/Events/IProjectEvent.ts
/* eslint-disable */
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import type {IFwEvent} from './IFwEvent';
import type {IProjectIdentifier} from '../../MiniLcm/Models/IProjectIdentifier';
import type {FwEventType} from './FwEventType';

export interface IProjectEvent extends IFwEvent
{
	event: IFwEvent;
	project: IProjectIdentifier;
	type: FwEventType;
	isGlobal: boolean;
}
/* eslint-enable */