// frontend/viewer/src/lib/dotnet-types/generated-types/FwLiteShared/Events/ISyncEvent.ts
/* eslint-disable */
//     This code was generated by a Reinforced.Typings tool.
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import type {IFwEvent} from './IFwEvent';
import type {FwEventType} from './FwEventType';
import type {SyncStatus} from '../../LexCore/Sync/SyncStatus';

export interface ISyncEvent extends IFwEvent
{
	type: FwEventType;
	isGlobal: boolean;
	status: SyncStatus;
}
/* eslint-enable */